1753-mcp-server/requirements.txt
mcp-server
requests>=2.31.0
